feat: redirecionando para tela de detalhes

+++ back-end/src/database/migrations/20230228190511-create-sale.js
@@ -31,7 +31,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.FLOAT(9, 2),
+        type: Sequelize.DECIMAL(10,2),
       },
       delivery_address: {
         allowNull: false,
