+++ back-end/src/database/models/SaleModel.js
@@ -13,7 +13,7 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.INTEGER, 
       foreignKey: true
     },
-    totalPrice: DataTypes.DOUBLE,
+    totalPrice: DataTypes.DECIMAL,
     deliveryAddress: DataTypes.STRING,
     deliveryNumber: DataTypes.STRING,
     saleDate: DataTypes.DATE,
